Clean up redux store initialization
